requirements.txt
llama-index
llama-index-llms-ollama
llama-index-embeddings-ollama
PyYAML
streamlit
watchdog